pyproject.toml
# Copyright 2023 Instadeep Inc
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "sebulba"
description = "InstaDeeps implementation of the Sebulba Distributed RL Architecture from DeepMind's Podracer Paper"
version = "0.0.1"
authors = [
    {name = "Instadeep Ltd", email = "hello@instadeep.com"}
]
requires-python = ">=3.9"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "chex",
    "envpool",
    "flax",
    "hydra-core",
    "numpy",
    "omegaconf",
    "optax",
    "requests",
    "rlax",
    "tensorboardx",
    "jax",
]

[project.urls]
Repository = "https://github.com/instadeepai/sebulba"

[project.optional-dependencies]
neptune = [
    "neptune"
]

[tool.setuptools.packages.find]
include = ["sebulba*"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.8
namespace_packages = true
incremental = false
cache_dir = ""
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false
allow_redefinition = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
strict_optional = true
strict_equality = true
explicit_package_bases = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = [
    "numpy.*",
    "pytest.*",
    "optax.*",
    "envpool.*",
    "neptune.*",
    "rlax.*",
    "tensorboardX.*",
]
ignore_missing_imports = true
